// types.ts
export enum CorruptionMethod {
  BLOCK = 'Block Out',
  LEET = '1337 Speak',
  GLITCH = 'Glitch',
  SCRAMBLE = 'Scramble',
  REVERSE = 'Reverse',
  VOWEL_EATER = 'Vowel Eater',
  CONSONANT_EATER = 'Consonant Eater',
  UNICODE_ABUSE = 'Unicode Abuse',
  CHARACTER_SWAP = 'Character Swap',
  EMOJI = 'Emoji',
}